Crate gluon_base

The base crate contains pervasive types used in the compiler such as type representations, the AST and some basic containers.

Modules

ast
Module containing the types which make up gluon's AST (Abstract Syntax Tree)

error
Module containing a few common error wrappers which allows more information to be saved for later display to the user

fixed
Collection types which allows insertion of new values while shared references to its contents are alive. This is done by storing each value in a stable memory location and preventing an earlier inserted value to be overwritten.

fnv

instantiate

kind

metadata

pos
Source code locations (borrowed from rustc's libsyntax_pos)

scoped_map
A map data type which allows the same key to exist at multiple scope levels

source
Module containing types and functions for mapping between byte indexes and line and column locations

symbol
Module which contains types working with symbols

types