Trait gluon_base::ast::Typed

pub trait Typed {
    type Ident;
    fn env_type_of(&self, env: &TypeEnv) -> ArcType<Self::Ident>;
}

Trait which abstracts over things that have a type. It is not guaranteed that the correct type is returned until after typechecking

Associated Types

type Ident

Required Methods

fn env_type_of(&self, env: &TypeEnv) -> ArcType<Self::Ident>