Module gluon_base::instantiate

Structs

Instantiator

Enums

Error

Functions

instantiate

maybe_remove_alias
Expand typ if it is an alias that can be expanded and return the expanded type. Returns None if the type is not an alias or the alias could not be expanded.

remove_alias

remove_aliases
Removes type aliases from typ until it is an actual type

remove_aliases_checked
Removes all possible aliases while checking that

remove_aliases_cow

type_of_alias
Returns the type which is aliased by alias if it was possible to do a substitution on the type arguments of alias using args