Module gluon_base::pos

Source code locations (borrowed from rustc's libsyntax_pos)

Structs

BytePos
A byte offset in a source string

Column
A 0-indexed column number, displayed externally as if it were offset from 1.

ExpansionId
An expansion identifier tracks wheter a span originated from a macro expansion or not

Line
A 0-indexed line number, displayed externally as if it were offset from 1.

Location
A location in a source file

Span
A span between two locations in a source file

Spanned

Constants

NO_EXPANSION

UNKNOWN_EXPANSION

Functions

span

spanned

spanned2