Module gluon_base::symbol

Module which contains types working with symbols

Structs

Components

Name

NameBuf

Symbol
A symbol uniquely identifies something regardless of its name and which module it originated from

SymbolModule
SymbolModule wraps a Symbols struct and adds a prefix to all symbols created by the symbol method. While this prefix does not affect the uniques of a Symbol in any way, it does make the origin of a symbol clearer when pretty printing it

SymbolRef

Symbols
Symbols is a bidirectional mapping between Symbols and their name as represented in a source file Used to make identifiers within a single module point to the same symbol