Struct gluon_base::symbol::Symbols

pub struct Symbols { /* fields omitted */ }

Symbols is a bidirectional mapping between Symbols and their name as represented in a source file Used to make identifiers within a single module point to the same symbol

Methods

impl Symbols

fn new() -> Symbols

fn symbol<N>(&mut self, name: N) -> Symbol where N: Into<NameBuf> + AsRef<Name>

Looks up the symbol for name or creates a new symbol if it does not exist