Enum gluon_base::types::Type

pub enum Type<Id, T = ArcType<Id>> {
    Hole,
    Opaque,
    Builtin(BuiltinType),
    App(T, AppVec<T>),
    Record(T),
    Variant(T),
    EmptyRow,
    ExtendRow {
        types: Vec<Field<Id, Alias<Id, T>>>,
        fields: Vec<Field<Id, T>>,
        rest: T,
    },
    Ident(Id),
    Variable(TypeVariable),
    Generic(Generic<Id>),
    Alias(AliasData<Id, T>),
}

The representation of gluon's types.

For efficency this enum is not stored directly but instead a pointer wrapper which derefs to Type is used to enable types to be shared. It is recommended to use the static functions on Type such as Type::app and Type::record when constructing types as those will construct the pointer wrapper directly.

Variants
Hole

An unbound type _, awaiting ascription.

Opaque

An opaque type

Builtin(BuiltinType)

A builtin type

App(T, AppVec<T>)

A type application with multiple arguments. For example, Map String Int would be represented as App(Map, [String, Int]).

Record(T)

Record constructor, of kind Row -> Type

Variant(T)

Variant constructor, of kind Row -> Type

EmptyRow

The empty row, of kind Row

ExtendRow

Row extension, of kind ... -> Row -> Row

Fields

types: Vec<Field<Id, Alias<Id, T>>>

The associated types of this record type

fields: Vec<Field<Id, T>>

The fields of this record type

rest: T

The rest of the row

Ident(Id)

An identifier type. These are created during parsing, but should all be resolved into Type::Aliases during type checking.

Identifiers are also sometimes used inside aliased types to avoid cycles in reference counted pointers. This is a bit of a wart at the moment and may cause spurious unification failures.

Variable(TypeVariable)

An unbound type variable that may be unified with other types. These will eventually be converted into Type::Generics during generalization.

Generic(Generic<Id>)

A variable that needs to be instantiated with a fresh type variable when the binding is refered to.

Alias(AliasData<Id, T>)
Methods

impl<Id, T> Type<Id, T> where T: From<Type<Id, T>>

fn hole() -> T

fn opaque() -> T

fn array(typ: T) -> T

fn app(id: T, args: AppVec<T>) -> T

fn variant(fields: Vec<Field<Id, T>>) -> T

fn poly_variant(fields: Vec<Field<Id, T>>, rest: T) -> T

fn record(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>) -> T

fn poly_record(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>, rest: T) -> T

fn extend_row(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>, rest: T) -> T

fn empty_row() -> T

fn function(args: Vec<T>, ret: T) -> T where T: Clone

fn generic(typ: Generic<Id>) -> T

fn builtin(typ: BuiltinType) -> T

fn variable(typ: TypeVariable) -> T

fn alias(name: Id, args: Vec<Generic<Id>>, typ: T) -> T

fn ident(id: Id) -> T

fn string() -> T

fn char() -> T

fn byte() -> T

fn int() -> T

fn float() -> T

fn unit() -> T

impl<Id, T> Type<Id, T> where T: Deref<Target=Type<Id, T>>

fn as_function(&self) -> Option<(&T, &T)>

fn as_alias(&self) -> Option<(&Id, &[T])>

impl<T> Type<Symbol, T> where T: Deref<Target=Type<Symbol, T>>

fn name(&self) -> Option<&SymbolRef>

Returns the name of self Example: Option a => Option Int => Int