Enum gluon_base::types::Type

pub enum Type<Id, T = ASTType<Id>> {
    App(T, T),
    Data(T, Vec<T>),
    Variants(Vec<(Id, T)>),
    Variable(TypeVariable),
    Generic(Generic<Id>),
    Function(Vec<T>, T),
    Builtin(BuiltinType),
    Array(T),
    Record {
        types: Vec<Field<Id, Alias<Id, T>>>,
        fields: Vec<Field<Id, T>>,
    },
    Id(Id),
    Alias(AliasData<Id, T>),
}

The representation of gluon's types.

For efficency this enum is not stored directly but instead a pointer wrapper which derefs to Type is used to enable types to be shared. It is recommended to use the static functions on Type such as Type::app and Type::record when constructing types as those will construct the pointer wrapper directly.

Variants

App

An application which applies the first argument on the second. Option Int would be represented as App(Option, Int)

Data

An application with multiple arguments

Variants

A variant type | A Int Float | B. The second element of the tuple is the function type which the constructor has which in the above example means that A's type is Int -> Float -> A and B's is B

Variable

Representation for type variables

Generic

Variant for "generic" variables. These occur in signatures as lowercase identifers a, b etc and are what unbound type variables are eventually made into.

Function

A function type <type> -> <type>

Builtin

A builtin type

Array

An array type Array T

Record

A record type

Fields

types

The associated types of this record type

fields

The fields of this record type

Id

An identifier type. Anything which is not a builting type.

Alias

Methods

impl<Id, T> Type<Id, T> where T: Deref<Target=Type<Id, T>>

fn is_uninitialized(&self) -> bool

fn kind(&self) -> RcKind

impl<Id, T> Type<Id, T> where T: From<Type<Id, T>>

fn app(l: T, r: T) -> T

fn array(typ: T) -> T

fn data(id: T, args: Vec<T>) -> T

fn variants(vs: Vec<(Id, T)>) -> T

fn record(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>) -> T

fn function(args: Vec<T>, ret: T) -> T

fn generic(typ: Generic<Id>) -> T

fn builtin(typ: BuiltinType) -> T

fn variable(typ: TypeVariable) -> T

fn alias(name: Id, args: Vec<Generic<Id>>, typ: T) -> T

fn id(id: Id) -> T

fn string() -> T

fn char() -> T

fn byte() -> T

fn int() -> T

fn float() -> T

fn unit() -> T

impl<Id, T> Type<Id, T> where T: Deref<Target=Type<Id, T>>

fn as_alias(&self) -> Option<(&Id, &[T])>