Enum gluon_base::types::Type

pub enum Type<Id, T = ArcType<Id>> {
    Hole,
    App(T, Vec<T>),
    Variants(Vec<(Id, T)>),
    Variable(TypeVariable),
    Generic(Generic<Id>),
    Builtin(BuiltinType),
    Record(T),
    EmptyRow,
    ExtendRow {
        types: Vec<Field<Id, Alias<Id, T>>>,
        fields: Vec<Field<Id, T>>,
        rest: T,
    },
    Ident(Id),
    Alias(AliasData<Id, T>),
}

The representation of gluon's types.

For efficency this enum is not stored directly but instead a pointer wrapper which derefs to Type is used to enable types to be shared. It is recommended to use the static functions on Type such as Type::app and Type::record when constructing types as those will construct the pointer wrapper directly.

Variants
Hole

An unbound type _, awaiting ascription.

App(T, Vec<T>)

An application with multiple arguments. Map String Int would be represented as App(Map, [String, Int])

Variants(Vec<(Id, T)>)

A variant type | A Int Float | B. The second element of the tuple is the function type which the constructor has which in the above example means that A's type is Int -> Float -> A and B's is B

Variable(TypeVariable)

Representation for type variables

Generic(Generic<Id>)

Variant for "generic" variables. These occur in signatures as lowercase identifers a, b etc and are what unbound type variables are eventually made into.

Builtin(BuiltinType)

A builtin type

Record(T)

A record type

EmptyRowExtendRow
Fields

types: Vec<Field<Id, Alias<Id, T>>>

The associated types of this record type

fields: Vec<Field<Id, T>>

The fields of this record type

rest: T

Ident(Id)

An identifier type. Anything that is not a builtin type.

Alias(AliasData<Id, T>)
Methods

impl<Id, T> Type<Id, T> where T: From<Type<Id, T>>

fn hole() -> T

fn array(typ: T) -> T

fn app(id: T, args: Vec<T>) -> T

fn variants(vs: Vec<(Id, T)>) -> T

fn record(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>) -> T

fn poly_record(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>, rest: T) -> T

fn extend_row(types: Vec<Field<Id, Alias<Id, T>>>, fields: Vec<Field<Id, T>>, rest: T) -> T

fn empty_row() -> T

fn function(args: Vec<T>, ret: T) -> T where T: Clone

fn generic(typ: Generic<Id>) -> T

fn builtin(typ: BuiltinType) -> T

fn variable(typ: TypeVariable) -> T

fn alias(name: Id, args: Vec<Generic<Id>>, typ: T) -> T

fn ident(id: Id) -> T

fn string() -> T

fn char() -> T

fn byte() -> T

fn int() -> T

fn float() -> T

fn unit() -> T

impl<Id, T> Type<Id, T> where T: Deref<Target=Type<Id, T>>

fn as_function(&self) -> Option<(&T, &T)>

fn as_alias(&self) -> Option<(&Id, &[T])>

impl<T> Type<Symbol, T> where T: Deref<Target=Type<Symbol, T>>

fn name(&self) -> Option<&SymbolRef>

Returns the name of self Example: Option a => Option Int => Int